Moved code to separate files without changes
